% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_json}
\alias{as_json}
\title{Convert an object to JSON}
\usage{
as_json(x, ...)
}
\arguments{
\item{x}{R object to be converted to JSON.}

\item{...}{additional arguments (except \code{auto_unbox} and \code{null}) passed to
\code{\link[jsonlite]{toJSON}}}
}
\value{
character of class "json"
}
\description{
This function effectively wraps \code{\link[jsonlite]{toJSON}} with two
hardcoded arguments:
}
\details{
\itemize{
\item \code{auto_unbox} is set to \code{TRUE}, such that singleton values are not
represented as lists.
\item \code{NULL} values are represented as JSON \code{null}s.
}
}
\keyword{internal}
