% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-lambda.R
\name{assert_lambda_is_setup}
\alias{assert_lambda_is_setup}
\title{Assert that the Lambda runtime has been set up}
\usage{
assert_lambda_is_setup()
}
\value{
TRUE
}
\description{
This package maintains a package-internal environment that, amongst other
uses, records if Lambda runtime has been set up via the
\verb{\link{setup_lambda}} or \verb{\link{start_lambda}} functions. This
function asserts that this setup has been completed and recorded, and
errors otherwise.
}
\keyword{internal}
