% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{assert_status_code_is_good}
\alias{assert_status_code_is_good}
\title{Check that the status code shows a success, and error otherwise}
\usage{
assert_status_code_is_good(status_code)
}
\arguments{
\item{status_code}{integer, usually returned by
\code{\link[httr]{status_code}}}
}
\value{
TRUE
}
\description{
Check that the status code shows a success, and error otherwise
}
\keyword{internal}
