% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handling.R
\name{condition}
\alias{condition}
\title{Define a condition (like an error) with support for HTTP status codes}
\usage{
condition(subclass, message, code = 500L, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{conditions returned by this function will be of the class
\code{c(subclass, "simpleError", "error", "condition")}.}

\item{message}{character message, such as an error description, to include
with the condition.}

\item{code}{HTTP status code to return (if applicable). Defaults to \code{500},
which is a generic "Internal Server Error". Not currently used anywhere.}

\item{call}{call expression}

\item{...attributes, }{specified in \code{tag = value} form, which will be attached
to the condition}
}
\description{
For more information on conditions and errors see
\url{http://adv-r.had.co.nz/Exceptions-Debugging.html}. See also
\code{\link[base]{conditions}}.
}
\keyword{internal}
