% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{config_or_runtime_api}
\alias{config_or_runtime_api}
\title{Convert a config to a runtime API if necessary}
\usage{
config_or_runtime_api(config)
}
\arguments{
\item{config}{A list of configuration values as created by the
\code{lambda_config} function.}
}
\description{
Endpoint functions need to accept either a config (as created by
\code{\link{lambda_config}}) or a runtime API. This function will accept
either and ensure that the runtime API is returned.
}
\keyword{internal}
