% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{decode_html_body}
\alias{decode_html_body}
\title{Decode the body of event content coming via an API Gateway}
\usage{
decode_html_body(body, config, base64_encoded = FALSE)
}
\arguments{
\item{body}{character body of an event received via an API Gateway
invocation. Usually this isn't the entire content of the event, but the
"body" component of it.}

\item{config}{A list of configuration values as created by the
\code{lambda_config} function.}

\item{base64_encoded}{logical that indicates if the body is encoded as Base64}
}
\value{
Either a list or, if the body is Base64 and the configuration demands
that Base64 values are not decoded, a Base64 value as a character
}
\description{
Decode the body of event content coming via an API Gateway
}
\keyword{internal}
