% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-context.R
\name{extract_context_from_environment}
\alias{extract_context_from_environment}
\title{Extract context from environment variables}
\usage{
extract_context_from_environment()
}
\value{
list
}
\description{
This function is intended to provide ambient configuration that makes up part
of the context returned by \code{\link{extract_context}}. These are the
components of the context formed by environment variables. To speed things up
a little we call the environment variables only once on runtime
initialisation and store them in the config.
}
\keyword{internal}
