% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{extract_event_headers}
\alias{extract_event_headers}
\title{Extract the headers from a Lambda event}
\usage{
extract_event_headers(event)
}
\description{
This function is largely equivalent to \code{\link[httr]{headers}}, which it
wraps. The only difference is that the names of the headers returned are
converted to lower-case (these are meant to be case-insensitive) and the
headers are logged at the DEBUG level.
}
\keyword{internal}
