% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-listening.R
\name{extract_request_id_from_headers}
\alias{extract_request_id_from_headers}
\title{Extract the request ID from the headers of an event, or error otherwise}
\usage{
extract_request_id_from_headers(headers)
}
\arguments{
\item{headers}{headers of a HTML response, as extracted by
\code{\link[httr]{headers}} or \code{\link{extract_event_headers}}}
}
\value{
character
}
\description{
The Request ID is unique for each input of a Lambda. It is carried by the
"lambda-runtime-aws-request-id" header of the response from the next
invocation endpoint (see \code{\link{endpoints}}).
}
\keyword{internal}
