% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_base64}
\alias{from_base64}
\title{Decode a Base64 encoded value to a string}
\usage{
from_base64(x)
}
\arguments{
\item{x}{a Base64 string}
}
\value{
character
}
\description{
Events coming via an API Gateway can have content with bodies encoded as
Base64. This is especially true for HTML API Gateways (as opposed to REST
API Gateways).
}
\examples{
from_base64("eyJudW1iZXIiOjd9")
}
