% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-config.R
\name{get_handler_function_from_env_var}
\alias{get_handler_function_from_env_var}
\title{Determine the function referred to by the "_HANDLER" environment variable}
\usage{
get_handler_function_from_env_var(environ)
}
\arguments{
\item{environ}{environment in which to search for the function given by the
"_HANDLER" environment variable. Defaults to the parent frame.}
}
\description{
This function will try to identify the function referred to by the "_HANDLER"
environment variable. This environment variable is configured by AWS Lambda,
either through the CMD of the Dockerfile containing the runtime or through
the AWS Lambda console (which takes priority). This function also performs
some checks, making sure that the environment variable is defined and that
it exists in the given environment, and that it's a function.
}
\keyword{internal}
