% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-setup.R
\name{get_lambda_runtime_api}
\alias{get_lambda_runtime_api}
\title{Retrieve the Lambda Runtime API value}
\usage{
get_lambda_runtime_api()
}
\value{
character
}
\description{
This value is set once per Lambda instance. It is used to construct the
various endpoints required to connect with the Lambda service.
}
\keyword{internal}
