% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handling.R
\name{handle_decomposition_error}
\alias{handle_decomposition_error}
\title{Handle an error that occurs when decomposing an invocation into an event}
\usage{
handle_decomposition_error(request_id)
}
\arguments{
\item{request_id}{For \code{get_response_endpoint} and
\code{get_invocation_error_endpoint}, the ID of the particular event/request. This
is provided in the "lambda-runtime-aws-request-id" header of the event.}
}
\value{
A function that accepts an error \code{e} as caught by
\code{\link[base]{tryCatch}}
}
\description{
This function is largely similar to \code{\link{handle_event_error}}. Its
purpose is to catch errors that occur between identifying the request ID
of an invocation and handling the event it defines. As such, it takes a
\code{request_id} argument rather than an \verb{event argument}.
}
\keyword{internal}
