% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{handle_event}
\alias{handle_event}
\title{Process the input of an event, and submit the result to Lambda}
\usage{
handle_event(event, deserialiser = NULL, serialiser = NULL)
}
\arguments{
\item{event}{the response received from querying the next invocation
endpoint.}

\item{deserialiser}{function for deserialising the body of the event.
By default, will attempt to deserialise the body as JSON, based on whether
the input is coming from an API Gateway, scheduled Cloudwatch event, or
direct. To use the body as is, pass the \code{identity} function.}

\item{serialiser}{function for serialising the result before sending.
By default, will attempt to serialise the body as JSON, based on the
request type. To send the result as is, pass the \code{identity} function.}
}
\description{
Process the input of an event, and submit the result to Lambda
}
\keyword{internal}
