% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handling.R
\name{handle_invocation_error}
\alias{handle_invocation_error}
\title{Generate a handling function for an invocation error}
\usage{
handle_invocation_error(event)
}
\arguments{
\item{e}{error caught by \code{\link[base]{tryCatch}}}
}
\description{
An error caught during event handling must be handled in a special way. An
error message must be logged and posted to the invocation error endpoint,
and the the runtime must continue --- an invocation error is a problem for
the invocation, not the runtime.

The \code{handle_invocation_error} function accepts an event and generates a
function. The generated function accepts error caught by
\code{\link[base]{tryCatch}}, logs it, and then submits it to the invocation
error endpoint. Importantly it does not stop the kernel --- the intention is
that the runtime moves onto the next event.\if{html}{\out{<div class="r">}}\preformatted{tryCatch(
    handle_event(...),
    error = handle_invocation_error(event) # returns a function(e)
 )
}\if{html}{\out{</div>}}
}
\keyword{internal}
