% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge-events.R
\name{is_eventbridge_event_content}
\alias{is_eventbridge_event_content}
\title{Determine if a Lambda event is coming from a EventBridge event (CloudWatch events)}
\usage{
is_eventbridge_event_content(event_content)
}
\arguments{
\item{event_content}{the content of the response received from querying the
text invocation endpoint, as a character}
}
\value{
logical
}
\description{
See \url{https://docs.aws.amazon.com/lambda/latest/dg/services-cloudwatchevents.html}
for more information.
}
\keyword{internal}
