% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{is_from_api_gateway}
\alias{is_from_api_gateway}
\title{Determine if a Lambda event is coming via an API Gateway}
\usage{
is_from_api_gateway(event_content)
}
\arguments{
\item{event_content}{}
}
\value{
logical
}
\description{
Determine if a Lambda event is coming via an API Gateway
}
\section{Invocations via an API Gateway}{

Events coming from an API Gateway need to be treated a little differently,
both in parsing the event content and in posting the results. Refer to
\code{vignette("api-gateway-invocations", package = "lambdr")} for details.
}

\keyword{internal}
