% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{is_scheduled_event_content}
\alias{is_scheduled_event_content}
\title{Determine if a Lambda event is coming from a scheduled Cloudwatch event}
\usage{
is_scheduled_event_content(event_content)
}
\arguments{
\item{event_content}{}
}
\value{
logical
}
\description{
Determine if a Lambda event is coming from a scheduled Cloudwatch event
}
\keyword{internal}
