% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns-events.R
\name{is_sns_event_content}
\alias{is_sns_event_content}
\title{Determine if a Lambda event is coming from SNS}
\usage{
is_sns_event_content(event_content)
}
\arguments{
\item{event_content}{}
}
\value{
logical
}
\description{
See \url{https://docs.aws.amazon.com/lambda/latest/dg/with-sns.html}
for more information.
}
\keyword{internal}
