% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdr.R
\docType{package}
\name{lambdr}
\alias{lambdr}
\title{lambdr: Create a Runtime for Serving Containerised R Functions on AWS Lambda}
\description{
This package provides an R runtime for the
\href{https://aws.amazon.com/lambda/}{\emph{AWS Lambda} serverless compute
service}. It is intended to be used to create containers that can run on \emph{AWS
Lambda}. \code{lambdr} provides the necessary functionality for handling the
various endpoints required for accepting new input and sending responses.

This package is \strong{unofficial}. Its creators are not affiliated with \emph{Amazon
Web Services}, nor is its content endorsed by \emph{Amazon Web Services}.
\emph{Lambda}, \emph{API Gateway}, \emph{EventBridge}, \emph{CloudWatch}, and \emph{SNS} are services
of \emph{Amazon Web Services}.

To see an example of how to use this package to create a runtime, refer to
\code{vignette("lambda-runtime-in-container", package = "lambdr")}.

The following invocation types have been implemented to some degree. Refer to
the vignettes or the package website for more information.

Alternatively, custom functions can be provided for parsing event content and
serialising results. Refer to \code{\link{lambda_config}} for more
information.
}
\section{Direct invocations}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}

\section{REST API Gateway invocations}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{vignette("api-gateway-invocations", package = "lambdr")}
}

\section{HTML API Gateway invocations}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{vignette("api-gateway-invocations", package = "lambdr")}
}

\section{EventBridge invocations}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{vignette("eventbridge-and-sns-invocations", package = "lambdr")}
}

\section{SNS invocations}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{vignette("eventbridge-and-sns-invocations", package = "lambdr")}
}

