% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mark_as_already_serialised}
\alias{mark_as_already_serialised}
\title{Give a value the "already_serialised = TRUE" attribute}
\usage{
mark_as_already_serialised(x)
}
\arguments{
\item{x}{any R object}
}
\value{
x with "already_serialised" attribute TRUE
}
\description{
Give a value the "already_serialised = TRUE" attribute
}
\keyword{internal}
