% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{parse_event_content}
\alias{parse_event_content}
\title{Parse the body of the Lambda event}
\usage{
parse_event_content(event, deserialiser = NULL)
}
\arguments{
\item{event}{the response received from querying the next invocation
endpoint.}

\item{deserialiser}{function for deserialising the body of the event.
By default, will attempt to deserialise the body as JSON, based on whether
the input is coming from an API Gateway, scheduled Cloudwatch event, or
direct. To use the body as is, pass the \code{identity} function.}
}
\value{
A list containing the "arguments" and "content_type", the latter of
which is either "HTML", "scheduled", or "direct". The content type may be
used in serialising the response to be sent back to Lambda.
}
\description{
Parse the body of the Lambda event
}
\keyword{internal}
