% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_json_or_empty}
\alias{parse_json_or_empty}
\title{Parse a JSON, but force a NULL or empty string to be interpreted as an empty list}
\usage{
parse_json_or_empty(json, ...)
}
\arguments{
\item{json}{character to be interpreted as a JSON}

\item{...}{additional arguments passed to \code{\link[jsonlite]{fromJSON}}}
}
\value{
list
}
\description{
Since \code{jsonlite::fromJSON(NULL)} and \code{jsonlite::fromJSON("")} return errors,
this function will force a NULL or empty string to be interpreted as
\code{list()}. Otherwise, the output of this function is identical to
\code{\link[jsonlite]{fromJSON}}.
}
\keyword{internal}
