% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{prettify_list}
\alias{prettify_list}
\title{Convert a list to a single character, preserving names}
\usage{
prettify_list(x)
}
\arguments{
\item{x}{Named list.}
}
\value{
character
}
\description{
Convert a list to a single character, preserving names
}
\examples{
prettify_list(list("a" = 1, "b" = 2, "c" = 3))
# "a=5, b=5, c=5"
}
\keyword{internal}
