% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-lambda.R
\name{reset_lambda}
\alias{reset_lambda}
\title{Undo the Lambda runtime setup}
\usage{
reset_lambda()
}
\description{
This function removes all detected variables and setup initiated by either
\verb{\link{setup_lambda}} or \verb{\link{start_lambda}}. It does not affect
the logging handlers setup by these functions.
}
\keyword{internal}
