% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-setup.R
\name{setup_lambda}
\alias{setup_lambda}
\title{Set up endpoints, variables and logging for AWS Lambda}
\usage{
setup_lambda()
}
\description{
Set up endpoints, variables and logging for AWS Lambda
}
\details{
As a rule of thumb, it takes longer to retrieve a value from an environment
variable than it does to retrieve a value from R. This is because retrieving
an environment variable requires a system call. Since the environment
variables do not change in a Lambda instance, we fetch them once and set them
to a package environment.
}
