% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-setup.R
\name{setup_logging}
\alias{setup_logging}
\title{Set up logging with the \code{logger} package}
\usage{
setup_logging(log_threshold = logger::INFO)
}
\arguments{
\item{log_threshold}{Threshold for recording and displaying log entries.
Defaults to \code{logger::INFO}. To debug a failing Lambda container, set this
to \code{logger::DEBUG} to log verbose information about how each request is
processed.}
}
\description{
Set up logging with the \code{logger} package
}
\keyword{internal}
