% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{start_listening}
\alias{start_listening}
\title{Start listening for events, and process them as they come}
\usage{
start_listening(deserialiser = NULL, serialiser = NULL, timeout_seconds = NULL)
}
\arguments{
\item{deserialiser}{function for deserialising the body of the event.
By default, will attempt to deserialise the body as JSON, based on whether
the input is coming from an API Gateway, scheduled Cloudwatch event, or
direct. To use the body as is, pass the \code{identity} function.}

\item{serialiser}{function for serialising the result before sending.
By default, will attempt to serialise the body as JSON, based on the
request type. To send the result as is, pass the \code{identity} function.}

\item{timeout_seconds}{If set, the function will stop listening for events
after this timeout. The timeout is checked between events, so this won't
interrupt the function while it is waiting for a new event. This argument
is provided for testing purposes, and shouldn't otherwise need to be set:
AWS should handle the shutdown of idle Lambda instances.}
}
\description{
Start listening for events, and process them as they come
}
\keyword{internal}
