% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handling.R
\name{stop_missing_request_id}
\alias{stop_missing_request_id}
\alias{handle_missing_request_id}
\title{Raise an error when there's a missing request ID}
\usage{
stop_missing_request_id(message)

handle_missing_request_id(e)
}
\arguments{
\item{message}{character message, such as an error description, to include
with the condition.}

\item{e}{error caught by \code{\link[base]{tryCatch}}}
}
\description{
A missing Request ID is an unusual error. It occurs after the runtime has
been initialised, and is \emph{technically} an invocation error. However it's
impossible to post the error to the invocation error endpoint as that
requires a Request ID.

We define a separate condition for it here, such that instead of \code{stop} we
can use \code{stop_missing_request_id}. This allows us to catch and treat the
error in a unique way. The \code{handle_missing_request_id} function accepts an
error caught by \code{\link[base]{tryCatch}}, logs it, and then returns
a standard error with a sensible message.
}
\keyword{internal}
