% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-config.R
\name{validate_lambda_config}
\alias{validate_lambda_config}
\title{Validate a Lambda config object}
\usage{
validate_lambda_config(config)
}
\arguments{
\item{config}{A list of configuration values as created by the
\code{lambda_config} function.}
}
\value{
TRUE
}
\description{
This function only verifies that an object has a "lambda_config" S3 class.
}
\keyword{internal}
