% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-listening.R
\name{wait_for_and_handle_event}
\alias{wait_for_and_handle_event}
\title{Wait for and handle event}
\usage{
wait_for_and_handle_event(deserialiser = deserialiser, serialiser = serialiser)
}
\arguments{
\item{deserialiser}{function for deserialising the body of the event.
By default, will attempt to deserialise the body as JSON, based on whether
the input is coming from an API Gateway, scheduled Cloudwatch event, or
direct. To use the body as is, pass the \code{identity} function. If input is
coming via an API Gateway this will require some complicated parsing (see
below).}

\item{serialiser}{function for serialising the result before sending.
By default, will attempt to serialise the body as JSON, based on the
request type. To send the result as is, pass the \code{identity} function.}
}
\value{
\code{NULL}
}
\description{
Combines \code{\link{wait_for_event}} and \code{\link{handle_event}} along
with
}
\keyword{internal}
