% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{extract_context}
\alias{extract_context}
\title{Extract the context of a Lambda invocation from the headers of an event}
\usage{
extract_context(event, ...)
}
\arguments{
\item{event}{the response received from querying the next invocation
endpoint.}
}
\value{
list
}
\description{
Extract the context of a Lambda invocation from the headers of an event
}
\section{Event context}{

The \emph{context} of an event is a list of metadata about the invocation. It is
derived from the headers of a next event invocation response. By default it
consists of:
\itemize{
\item \code{aws_request_id} - The identifier of the invocation request
\item \code{invoked_function_arn} – The Amazon Resource Name (ARN) that's used to
invoke the function. Indicates if the invoker specified a version number or
alias.
}

Alternatively, a particular event \emph{class} (determined by invocation method)
can implement an \code{extract_context} method. This is useful for, say,
invocations coming via API Gateway, in which the context includes details
about the HTTP request. In all cases the context should be a list.

If the handler function accepts a \code{context} argument then it will
automatically receive at runtime a named list consisting of these values
along with the arguments in the body (if any). For example, a function such
as \code{my_func(x, context)} will receive the context argument automatically.
The \code{context} argument must be named (\code{...} will not work).
}

\keyword{internal}
