% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\name{generate_result}
\alias{generate_result}
\title{Parse the content of an event and pass it through the handler function}
\usage{
generate_result(event, config = lambda_config(), deserialiser = NULL)
}
\arguments{
\item{event}{the response received from querying the next invocation
endpoint.}

\item{config}{A list of configuration values as created by the
\code{lambda_config} function.}
}
\value{
An object of class the same as \code{event}. The object contains a
\code{result} value, and the \code{result_calculated} attribute is set to \code{TRUE}.
}
\description{
Parse the content of an event and pass it through the handler function
}
\keyword{internal}
