% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-event.R
\docType{data}
\name{is_from_api_gateway}
\alias{is_from_api_gateway}
\alias{http_request_element}
\title{Determine if a Lambda event is coming via an API Gateway}
\format{
An object of class \code{character} of length 1.
}
\usage{
is_from_api_gateway(event_content)

http_request_element
}
\arguments{
\item{event_content}{}
}
\value{
logical
}
\description{
Events coming from an API Gateway need to be treated a little differently,
both in parsing the event content and in posting the results. An event
coming from an API Gateway will contain the \code{http_request_element}.
}
\keyword{internal}
