% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdr.R
\docType{package}
\name{lambdr}
\alias{lambdr}
\title{lambdr: Create a Runtime for Serving Containerised R Functions on AWS Lambda}
\description{
Runtime for serving containers that can execute R code on the
\href{https://aws.amazon.com/lambda/}{\emph{AWS Lambda} serverless compute
service}. Provides the necessary functionality for handling the various
endpoints required for accepting new input and sending responses.

To see an example of how to use this package to create a runtime, refer to
\code{vignette("lambda-runtime-in-container", package = "lambdr")}.
}
\section{Direct invocations}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}

\section{API Gateway invocations}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}

\section{Cloudwatch scheduled event invocations}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}

