% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-event-content.R
\name{parse_event_content}
\alias{parse_event_content}
\title{Parse the body of the Lambda event}
\usage{
parse_event_content(event, deserialiser = NULL)
}
\arguments{
\item{deserialiser}{function for deserialising the body of the event.
By default, will attempt to deserialise the body as JSON, based on whether
the input is coming from an API Gateway, scheduled Cloudwatch event, or
direct. To use the body as is, pass the \code{identity} function. If input is
coming via an API Gateway this will require some complicated parsing (see
below).}

\item{event_content}{the content of the response received from querying the
text invocation endpoint, as a character}
}
\value{
A list containing the arguments to be passed to the handler function
}
\description{
Parse the body of the Lambda event
}
\section{Invocations via an API Gateway}{

Events coming from an API Gateway need to be treated a little differently,
both in parsing the event content and in posting the results. Refer to
\code{vignette("api-gateway-invocations", package = "lambdr")} for details.
}

\keyword{internal}
