% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-event-content.R
\name{parse_event_content}
\alias{parse_event_content}
\title{Parse the body of the Lambda event}
\usage{
parse_event_content(event, config)
}
\arguments{
\item{event}{the response received from querying the next invocation
endpoint.}

\item{config}{A list of configuration values as created by the
\code{lambda_config} function.}
}
\value{
A list containing the arguments to be passed to the handler function
}
\description{
Parse the body of the Lambda event
}
\section{Invocations via an API Gateway}{

Events coming from an API Gateway need to be treated a little differently,
both in parsing the event content and in posting the results. Refer to
\code{vignette("api-gateway-invocations", package = "lambdr")} for details.
}

\keyword{internal}
