% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-result.R
\name{post_result}
\alias{post_result}
\title{Post an event with a result to the response endpoint}
\usage{
post_result(event, config)
}
\arguments{
\item{event}{the response received from querying the next invocation
endpoint.}

\item{config}{A list of configuration values as created by the
\code{lambda_config} function.}
}
\description{
This function will first serialise the event result according to its class
by dispatching through \code{\link{serialise_result}}.The result of that
serialisation is posted \strong{as is} to the response endpoint; this function
will not perform any JSON serialisation, for example.
}
\section{Invocations via an API Gateway}{

Events coming from an API Gateway need to be treated a little differently,
both in parsing the event content and in posting the results. Refer to
\code{vignette("api-gateway-invocations", package = "lambdr")} for details.
}

\keyword{internal}
