% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-result.R
\name{serialise_result}
\alias{serialise_result}
\title{Serialise a result}
\usage{
serialise_result(event, config)
}
\arguments{
\item{event}{the response received from querying the next invocation
endpoint.}

\item{config}{A list of configuration values as created by the
\code{lambda_config} function.}
}
\value{
character.
}
\description{
This result is posted \strong{as is}, and so all JSON serialisation, etc. must be
performed here.
}
\section{Invocations via an API Gateway}{

Events coming from an API Gateway need to be treated a little differently,
both in parsing the event content and in posting the results. Refer to
\code{vignette("api-gateway-invocations", package = "lambdr")} for details.
}

\keyword{internal}
