% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-lambda.R
\name{setup_lambda}
\alias{setup_lambda}
\title{Set up endpoints, variables and logging for AWS Lambda}
\usage{
setup_lambda(
  handler = NULL,
  runtime_api = NULL,
  task_root = NULL,
  environ = parent.frame()
)
}
\arguments{
\item{handler}{character. Name of function to use for processing inputs from
events. This argument is provided for debugging and testing only. The
"_HANDLER" environment variable, as configured in AWS, as configured by
AWS, will always override this value if present.}

\item{runtime_api}{character. Used as the host in the various endpoints used
by AWS Lambda. This argument is provided for debugging and testing only.
The "AWS_LAMBDA_RUNTIME_API" environment variable, as configured by AWS,
will always override this value if present.}

\item{task_root}{character. Defines the path to the Lambda function code.
This argument is provided for debugging and testing only. The
"LAMBDA_TASK_ROOT" environment variable, as configured by AWS, will always
override this value if present.}

\item{environ}{environment in which to search for the function given by the
handler.}
}
\description{
Set up endpoints, variables and logging for AWS Lambda
}
\details{
As a rule of thumb, it takes longer to retrieve a value from an environment
variable than it does to retrieve a value from R. This is because retrieving
an environment variable requires a system call. Since the environment
variables do not change in a Lambda instance, we fetch them once and set them
to a package environment.
}
\section{AWS Lambda variables}{


The following variables are made available through environment variables
configured by AWS:
\itemize{
\item Lambda Runtime API, available as the "AWS_LAMBDA_RUNTIME_API" environment
variable, is the host of the various HTTP endpoints through which the
runtime interacts with Lambda. See also: \verb{\link{endpoints}}.
\item Lambda Task Root, available as the "LAMBDA_TASK_ROOT" environment variable,
defines the path to the Lambda function code.
\item The handler, available as the "_HANDLER" environment variable, is
interpreted by R as the function that is executed when the Lambda is
called. This value could be anything, as the interpretation is solely up
to the runtime, so requiring it to be a function is a standard imposed by
this package.
}
}

