% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-setup.R
\name{setup_lambda}
\alias{setup_lambda}
\title{Set up endpoints, variables and logging for AWS Lambda}
\usage{
setup_lambda(handler = NULL, log_threshold = logger::INFO)
}
\arguments{
\item{log_threshold}{Threshold for recording and displaying log entries.
Passed to logger::log_threshold. Defaults to \code{logger::INFO}. To debug a
failing Lambda container, set this to \code{logger::DEBUG} to log verbose
information about how each request is processed.}

\item{character.}{Name of function to use for processing inputs from events.
This argument is provided for debugging and testing only. The handler as
configured in AWS, if present, will always overwrite this value.}
}
\description{
Set up endpoints, variables and logging for AWS Lambda
}
\details{
As a rule of thumb, it takes longer to retrieve a value from an environment
variable than it does to retrieve a value from R. This is because retrieving
an environment variable requires a system call. Since the environment
variables do not change in a Lambda instance, we fetch them once and set them
to a package environment.
}
