% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-listening.R
\name{start_listening}
\alias{start_listening}
\title{Start listening for events, and process them as they come}
\usage{
start_listening(config = lambda_config(), timeout_seconds = NULL)
}
\arguments{
\item{config}{A list of configuration values as created by the
\code{lambda_config} function.}

\item{timeout_seconds}{If set, the function will stop listening for events
after this timeout. The timeout is checked between events, so this won't
interrupt the function while it is waiting for a new event. This argument
is provided for testing purposes, and shouldn't otherwise need to be set:
AWS should handle the shutdown of idle Lambda instances.}
}
\description{
Start listening for events, and process them as they come
}
\keyword{internal}
