% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-listening.R
\name{wait_for_and_handle_event}
\alias{wait_for_and_handle_event}
\title{Wait for and handle event}
\usage{
wait_for_and_handle_event(config)
}
\arguments{
\item{config}{A list of configuration values as created by the
\code{lambda_config} function.}
}
\value{
\code{NULL}
}
\description{
Combines \code{\link{wait_for_event}} and \code{\link{handle_event}} along
with
}
\keyword{internal}
